+++ src/main/java/org/mapdb/TxMaker.java
@@ -10,21 +10,33 @@ public class TxMaker {
 
     /** marker for deleted records*/
     protected static final Object DELETED = new Object();
+    private final boolean txSnapshotsEnabled;
+    private final boolean strictDBGet;
 
     /** parent engine under which modifications are stored */
-    protected org.mapdb.TxEngine engine;
+    protected Engine engine;
 
+    public TxMaker(Engine engine) {
+        this(engine,false,false);
+    }
 
-    public TxMaker(org.mapdb.TxEngine engine) {
+    public TxMaker(Engine engine, boolean strictDBGet, boolean txSnapshotsEnabled) {
         if(engine==null) throw new IllegalArgumentException();
-        if(engine.isReadOnly()) throw new IllegalArgumentException("read only");
-        if(!engine.canRollback()) throw new IllegalArgumentException("no rollback");
+        if(!engine.canSnapshot())
+            throw new IllegalArgumentException("Snapshot must be enabled for TxMaker");
+        if(engine.isReadOnly())
+            throw new IllegalArgumentException("TxMaker can not be used with read-only Engine");
         this.engine = engine;
+        this.strictDBGet = strictDBGet;
+        this.txSnapshotsEnabled = txSnapshotsEnabled;
     }
 
     
     public DB makeTx(){
-        return new DB(engine.snapshot());
+        Engine snapshot = engine.snapshot();
+        if(txSnapshotsEnabled)
+            snapshot = new TxEngine(snapshot,false);
+        return new DB(snapshot,strictDBGet);
     }
 
     public void close() {