+++ src/test/java/org/mapdb/HTreeMap3Test.java
@@ -42,7 +42,7 @@ public class HTreeMap3Test extends ConcurrentMapInterfaceTest<Integer, String> {
 
     @Override
     protected ConcurrentMap<Integer, String> makeEmptyMap() throws UnsupportedOperationException {
-        return new HTreeMap(r,0,0, HTreeMap.preallocateSegments(r), Serializer.BASIC, Serializer.BASIC,0,0,0,0,null,null,null,null);
+        return new HTreeMap(r,0,0, HTreeMap.preallocateSegments(r), Serializer.BASIC, Serializer.BASIC,0,0,0,0,0,null,null,null,null);
     }
 
     @Override